refactor(ui): share the Usage month plumbing between surfaces
The sandbox Usage section needs the same month stepper and calendar logic.
fillMonthDays carries the subtle part — zero-filling the server's sparse rows
and stopping at today for the current month — which a second copy would drift
from.
